[frontend] split refacto SubType routing (#14046)
