+++ opencti-platform/opencti-front/src/private/components/settings/sub_types/Root.tsx
@@ -5,15 +5,25 @@ import Loader from '../../../../components/Loader';
 import ErrorNotFound from '../../../../components/ErrorNotFound';
 import SubType from './SubType';
 import FintelTemplate from './fintel_templates/FintelTemplate';
+import SubTypeOverview from './SubTypeOverview';
+import SubTypeWorkflow from './SubTypeWorkflow';
+import useHelper from '../../../../utils/hooks/useHelper';
 
 const RootSubType = () => {
   const { subTypeId } = useParams<{ subTypeId?: string }>();
+
+  const { isFeatureEnable } = useHelper();
+  const isDraftWorkflowFeatureEnabled = isFeatureEnable('DRAFT_WORKFLOW');
+
   if (!subTypeId) return <ErrorNotFound />;
 
   return (
     <Suspense fallback={<Loader />}>
       <Routes>
-        <Route path="/" element={<SubType />} />
+        <Route path="/" element={<SubType />}>
+          <Route index element={<SubTypeOverview />} />
+          {isDraftWorkflowFeatureEnabled && <Route path="workflow" element={<SubTypeWorkflow />} />}
+        </Route>
         <Route
           path="/templates/:templateId"
           element={(
